src/main/java/com/ponto/inteligente/api/utils/PasswordUtils.java
package com.ponto.inteligente.api.utils;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;

public class PasswordUtils {
	
	private static final Logger log = LoggerFactory.getLogger(PasswordUtils.class);
	
	public static String generateCrypt(final String password) {
		if (password == null) {
			return password;
		}
		
		log.info("Generating password with bcrypt");
		
		BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder();
		return bCryptPasswordEncoder.encode(password);
		
	}

}
